+++ database.js
@@ -10,6 +10,7 @@ const mysql = require('mysql2');
 
 //Import models
 const User = require('./models/userModel');
+const School = require('./models/schoolModel')
 
 //Connect to database
 const sequelize = new Sequelize('FacultyElectionsDB', 'root', 'example', {
@@ -26,4 +27,5 @@ sequelize.authenticate().then(() => {
 module.exports = {
     // Gets the users with given options
     getUsers: async (options = {}) => User.findAll(options),
+    getSchools: async (options = {}) => School.findAll(options)
 }
